latest update

+++ assignment_2/part2/gaussDer.m
@@ -12,7 +12,7 @@ values = (-temp):1:(temp) ;
 constant = -1/(sigma^2) ;
 
 for i=1:length(G)
-    Gd(i) = constant * values(i) * G(i);
+    Gd(i) = double(constant) * double(values(i)) * G(i);
 end
 
 %Do convolution
